requirements.txt
amqp==2.6.1
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.0
cssselect2==0.3.0
defusedxml==0.7.0rc1
Django==2.2.16
django-braces==1.14.0
django-ckeditor==5.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.8
django-js-asset==1.2.2
django-markdown2==0.3.1
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-parler==2.2
django-phonenumber-field==4.0.0
django-rosetta==0.8.1
django-star-ratings==0.9.1
django-taggit==1.3.0
djangorestframework==3.11.1
docutils==0.16
flower==0.9.2
html5lib==1.1
humanize==3.0.1
idna==2.10
kombu==4.6.11
markdown2==2.3.9
microsofttranslator==0.8
oauthlib==3.1.0
pdfrw==0.4
phonenumbers==8.12.9
Pillow==7.2.0
pip-autoremove==0.9.1
polib==1.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
swapper==1.1.2.post1
tinycss2==1.0.2
tornado==4.2
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.16.0
